Bug 1453675 - Refactor ping structure
This splits pings' package into multiple files, one per ping type.
Information about presence of certain subobjects is encoded in ping types using HasEnvironment and HasApplication traits, which contain extractor methods with implementations specific to these objects.

With this change it was also possible to generify `ErrorAggregator#buildDimensions` to handle all ping types.
